[Patmos] Make recursive resolution of XForm patterns optional.
This fixes the failing tests in the CellSPU backend. With this option enabled,
the ISEL matcher for the CellSPU replaces a BUILD_VECTOR with a TargetContant
node in some cases, causing an assertion in some tests. Without this option
enabled, some Patmos patterns like SHADD2 fail to match properly and the
backend creates incorrect code.