[Patmos] Instruction printing no longer requires modifier.
Printing previously used the 'addrmod' operand modifier in
tablegen to signal that a given operand was part of an address.
E.g. that the 'r1' and '3' in '[r1 + 3]' should be printed in
a specific way.

This has been replaced with simply checking for which instruction
is being printed and which operand, which gives the same result.