[Patmos] Moved MachineModulePass into the Patmos folder.
MachineModulePass previously resided in llvm/CodeGen. This means
during the port to the new LLVM version, it was not included.
This commit copies MachineModulePass from the previous llvm Patmos port's
llvm/CodeGen folders (include and lib) directly into the Patmos folder,
as this class is a customization on of the previous llvm port and therefore
not used by other targets.